backend/pyproject.toml
[project]
name = "jesse-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "flask>=3.0.0",
  "flask-cors>=4.0.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "gunicorn>=22.0.0",
  "thefuzz>=0.22.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
